requirements.txt
numpy>=1.24
matplotlib>=3.7
pytest>=7.0
pyserial>=3.5
scipy>=1.10
